added/middleware/updated routes/added auth

+++ controllers/clothingItem.js
@@ -1,5 +1,10 @@
 const ClothingItem = require("../models/clothingItem");
-const { BAD_REQUEST, NOT_FOUND, SERVER_ERROR } = require("../utils/errors");
+const {
+  BAD_REQUEST,
+  NOT_FOUND,
+  SERVER_ERROR,
+  FORBIDEN,
+} = require("../utils/errors");
 
 const createItem = (req, res) => {
   const { name, weather, imageUrl } = req.body;
@@ -39,10 +44,22 @@ const getItems = (req, res) => {
 const deleteItem = (req, res) => {
   const { itemId } = req.params;
 
-  ClothingItem.findByIdAndDelete(itemId)
-    .orFail()
+  ClothingItem.findById(itemId)
+    .orFail(() => {
+      const error = new Error("Item not found");
+      error.name = "DocumentNotFoundError";
+      throw error;
+    })
     .then((item) => {
-      res.status(200).send(item);
+      if (item.owner.toString() !== userId) {
+        return res
+          .status(FORBIDEN)
+          .send({ message: "Not allowed to delete Item" });
+      }
+      return ClothingItem.findByIdAndDelete(itemId);
+    })
+    .then((deletedItem) => {
+      res.status(200).send({ message: "Item was deleted", deletedItem });
     })
     .catch((err) => {
       if (err.name === "DocumentNotFoundError") {
